fix border cases

+++ psa_soporte/employeeService.py
@@ -3,7 +3,7 @@ from .models import *
 import requests
 
 
-url = 'https://tpg-prueba.herokuapp.com/Recursos/'
+url = 'https://anypoint.mulesoft.com/mocking/api/v1/sources/exchange/assets/754f50e8-20d8-4223-bbdc-56d50131d0ae/recursos-psa/1.0.0/m/api/recursos'
 
 
 class EmployeeService:
@@ -47,6 +47,10 @@ class EmployeeService:
 
     def removeEmployeeFromTicket(self, employeeID, ticketID):
         db: Session = SessionLocal()
+
+        exists = db.query(Employee).filter_by(ticketID=ticketID, employeeID=employeeID).first() is not None
+        if not exists:
+            raise Exception('Cannot remove an employee that is not assigned to that ticket')
 
         db.query(Employee).filter_by(employeeID=employeeID, ticketID=ticketID).delete()
